Aligned storage of 'path' type.
Related change is that container of namespace declarations now points to offsets in the binary parser output as opposed to pointers. The pointers broke the parser as soon as reallocation took place due to increased volume of the aligned paths inside XNT. It was fixed in this commit to keep regression tests working.